pyproject.toml
[project]
name = "capstone-laskarai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.12"
dependencies = [
    "numpy>=2.1.3",
    "pillow>=11.2.1",
    "streamlit>=1.45.1",
    "tensorflow>=2.19.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
